fix(routes): include email module inside auth module
