+++ src/modules/auth/dto/reconfirm-doctor.dto.ts
@@ -1,15 +1,5 @@
-import {
-  IsEmail,
-  IsNotEmpty,
-  IsString,
-  IsDateString,
-  IsEnum,
-  MinLength,
-  Matches,
-  IsInt,
-} from 'class-validator';
+import { IsEmail, IsNotEmpty } from 'class-validator';
 import { ApiProperty } from '@nestjs/swagger';
-import { Role } from '../../../common/enums/Role';
 
 export class ReconfirmDoctorDto {
   @ApiProperty({
