+++ src/modules/auth/templates/welcome.pug
