--- src/modules/email/email.controller.ts
@@ -1,12 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-import { EmailService } from './email.service';
-
-@Controller('email')
-export class EmailController {
-  constructor(private readonly emailService: EmailService) {}
-
-  @Get()
-  async sendMail(to, link) {
-    return this.emailService.sendConfirmationLink(to, link);
-  }
-}
