--- src/modules/email/email.module.ts
@@ -1,35 +1,0 @@
-import { Module } from '@nestjs/common';
-import { MailerModule } from '@nestjs-modules/mailer';
-import { PugAdapter } from '@nestjs-modules/mailer/dist/adapters/pug.adapter';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { EmailService } from './email.service';
-import { EmailController } from './email.controller';
-
-@Module({
-  imports: [
-    MailerModule.forRootAsync({
-      imports: [ConfigModule],
-      useFactory: (configService: ConfigService) => ({
-        transport: {
-          host: configService.get('SMTP_SERVER'),
-          auth: {
-            user: configService.get('SMTP_LOGIN'),
-            pass: configService.get('SMTP_PASS'),
-          },
-        },
-        template: {
-          dir: `${__dirname}/templates`,
-          adapter: new PugAdapter(),
-          options: {
-            strict: true,
-          },
-        },
-      }),
-      inject: [ConfigService],
-    }),
-  ],
-  controllers: [EmailController],
-  providers: [EmailService],
-  exports: [EmailService],
-})
-export class EmailModule {}
